+++ src/v3/server.rs
@@ -270,7 +270,7 @@ fn handshake_service_factory<Io, St, C>(
 ) -> impl ServiceFactory<
     Config = (),
     Request = Io,
-    Response = (Io, IoState<mqtt::Codec>, Session<St>, usize),
+    Response = (Io, IoState<mqtt::Codec>, Session<St>, u16),
     Error = MqttError<C::Error>,
 >
 where
@@ -305,7 +305,7 @@ fn handshake_service_factory2<Io, St, C>(
 ) -> impl ServiceFactory<
     Config = (),
     Request = (Io, IoState<mqtt::Codec>),
-    Response = (Io, IoState<mqtt::Codec>, Session<St>, usize),
+    Response = (Io, IoState<mqtt::Codec>, Session<St>, u16),
     Error = MqttError<C::Error>,
     InitError = C::InitError,
 >
@@ -339,7 +339,7 @@ async fn handshake<Io, S, St, E>(
     service: S,
     max_size: u32,
     pool: Rc<MqttSinkPool>,
-) -> Result<(Io, IoState<mqtt::Codec>, Session<St>, usize), S::Error>
+) -> Result<(Io, IoState<mqtt::Codec>, Session<St>, u16), S::Error>
 where
     Io: AsyncRead + AsyncWrite + Unpin,
     S: Service<Request = Connect<Io>, Response = ConnectAck<Io, St>, Error = MqttError<E>>,
@@ -391,12 +391,7 @@ where
                         )
                         .await?;
 
-                    Ok((
-                        ack.io,
-                        ack.state,
-                        Session::new(session, ack.sink),
-                        ack.keepalive as usize,
-                    ))
+                    Ok((ack.io, ack.state, Session::new(session, ack.sink), ack.keepalive))
                 }
                 None => {
                     log::trace!(
